Split inputs into API and implementation files
