+++ src/neural_data_simulator/core/inputs/__init__.py
@@ -1,0 +1,1 @@
+"""Inputs API and implementation."""
