+++ src/neural_data_simulator/scripts/run_ephys_generator.py
@@ -17,7 +17,6 @@ from typing import cast, Optional
 
 import numpy as np
 
-from neural_data_simulator.core import inputs
 from neural_data_simulator.core import outputs
 from neural_data_simulator.core.ephys_generator import ContinuousData
 from neural_data_simulator.core.ephys_generator import LSLSpikeRateInputAdapter
@@ -28,6 +27,7 @@ from neural_data_simulator.core.ephys_generator import SpikeRateTestingInput
 from neural_data_simulator.core.ephys_generator import Spikes
 from neural_data_simulator.core.ephys_generator import Waveforms
 from neural_data_simulator.core.health_checker import HealthChecker
+from neural_data_simulator.core.inputs.lsl_input import LSLInput
 from neural_data_simulator.core.outputs import StreamConfig
 from neural_data_simulator.core.settings import EphysGeneratorEndpointType
 from neural_data_simulator.core.settings import EphysGeneratorSettings
@@ -72,7 +72,7 @@ def _setup_LSL_input(
     Returns:
         LSL stream input that can be used to read data from.
     """
-    lsl_inlet = inputs.LSLInput(stream_name, connection_timeout)
+    lsl_inlet = LSLInput(stream_name, connection_timeout)
     spike_rate_input = LSLSpikeRateInputAdapter(lsl_inlet)
     return spike_rate_input
 
@@ -232,7 +232,7 @@ def run():
         _get_spikes_params(settings.ephys_generator),
     )
 
-    outputs = ProcessOutput.LSLOutputs(
+    outputs = ProcessOutput.Outputs(
         continuous_data_output, lfp_output, spike_events_output
     )
