+++ src/neural_data_simulator/util/runtime.py
@@ -10,7 +10,7 @@ import sys
 from types import ModuleType
 from typing import Any, Optional, Union
 
-from neural_data_simulator.core.inputs import Input
+from neural_data_simulator.core.inputs.api import Input
 from neural_data_simulator.core.outputs import Output
 from neural_data_simulator.core.settings import LogLevel
